+++ sdk.js
@@ -298,6 +298,7 @@ class PaddleSDK {
 	getSubscriptionTransactions(subscriptionID, page) {
 		return this._getTransactions('subscription', subscriptionID, page);
 	}
+
 	/**
 	 * Get the list of transations for an order
 	 *
@@ -422,7 +423,7 @@ class PaddleSDK {
 	 *
 	 * @method
 	 * @param {number} subscriptionID
-	 * @param {Object} postData { quantity, price, planID, currency, prorate, keepModifiers, billImmediately }
+	 * @param {Object} postData { quantity, price, planID, currency, prorate, keepModifiers, billImmediately, pause }
 	 * @returns {Promise}
 	 * @fulfill {object} - The result of the operation
 	 *
@@ -438,6 +439,7 @@ class PaddleSDK {
 			prorate,
 			keepModifiers,
 			billImmediately,
+			pause,
 		} = postData;
 		const body = {
 			subscription_id: subscriptionID,
@@ -462,6 +464,9 @@ class PaddleSDK {
 		}
 		if (prorate) {
 			body.prorate = prorate;
+		}
+		if (typeof pause === 'boolean') {
+			body.pause = pause;
 		}
 
 		return this._request('/subscription/users/update', {
@@ -584,6 +589,38 @@ class PaddleSDK {
 	getOrderDetails(checkoutId) {
 		return this._request(`/order?checkout_id=${checkoutId}`, {
 			checkoutAPI: true,
+		});
+	}
+
+	/**
+	 * Create a subscription modifier to dynamically change the subscription payment amount
+	 *
+	 * @method
+	 * @param subscriptionID
+	 * @param modifierAmount
+	 * @param options
+	 * @returns {Promise}
+	 * @fulfil {object} - The result of the operation
+	 *
+	 * @example
+	 * const result = await client.createSubscriptionModifier(123, 10);
+	 * const result = await client.createSubscriptionModifier(123, 10, { modifier_recurring: false, modifier_description: 'description' });
+	 */
+	createSubscriptionModifier(subscriptionID, modifierAmount, options = {}) {
+		const { modifier_recurring, modifier_description } = options;
+		const body = {
+			subscription_id: subscriptionID,
+			modifier_amount: modifierAmount,
+		};
+		if (typeof modifier_recurring === 'boolean') {
+			body.modifier_recurring = modifier_recurring;
+		}
+		if (modifier_description) {
+			body.modifier_description = modifier_description;
+		}
+
+		return this._request('/subscription/modifiers/create', {
+			body: body,
 		});
 	}
 }